backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "email-validator>=2.2.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.0",
    "openpyxl>=3.1.5",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "qrcode[pil]>=8.0",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.34.0",
]
